requirements.txt
flask==3.0.0
--extra-index-url https://download.pytorch.org/whl/cu118
torch==2.2.0+cu118
torchvision==0.17.0+cu118
opencv-python==4.8.1.78
pillow==10.1.0
numpy==1.24.3
basicsr>=1.4.2
facexlib>=0.2.5
gfpgan>=1.3.5
realesrgan
